Work around quopri change in Python 3
qopri.encodestring() only accepts bytes objects and encodestring()
and decodestring() both output bytes objects in Python 3. To maintain
compatibility with Python 2, this requires encoding the input to
encodestring to bytes (a no-op on Python 2) and then checking to see if
we need to decode on the other side (which will depend on the version of
Python).