Make script compatible with Rocky 8 and CentOS 8 and add OS specific tags
A mixture of inconsistent testing and dependencies led to some weird errors that were hard to reproduce.
I cordoned it down to:
- Ordering of --depends calls, where calling python3-pip before python3-stomppy led to an almost instant
fail
- Bad module names
It was not in fact the ifs that were at fault.  This script has now been tested with Rocky 8 and CentOS 8
from start to finish, and Debian previously.  It will now be tested with EL7.