Add python3-dirq as dependency for apel-ssm-service
The receiver has to use dirq when receiving as that's what the APEL
loader expects to find. Whereas the non-dirq sending is a convenience
for non-APEL client sites who just want to dump files into a directory
for sending without worrying about the dirq naming schema (so it is
optional).