Remove default host and tidy comments
We don't want people accidentally sending to the production host, but we
also don't want people to accidentally send to the devel host if not
intended. So remove the default host and add the list of hosts to a
comment, explaining the purpose of each.

Also uncomment the port option and clarify its usage. It does nothing
for AMS sending, so might as well leave it uncommented.